move getCore to utils
